trac-ui/requirements.txt
crispy-bootstrap5
django
django-crispy-forms
djangorestframework
docker
nbconvert
nbformat
